+++ Joker.Kafka/KSql/Linq/IAggregations.cs
@@ -5,11 +5,18 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
   public interface IAggregations
   { 
     int Count();
-    // long LongCount(); //TOOD:
+    long LongCount();
   }
 
   public interface IAggregations<out TSource> : IAggregations
   {
+    short Count(Func<TSource, short> selector);
+    int Count(Func<TSource, int> selector);
+    long Count(Func<TSource, long> selector);
+    float Count(Func<TSource, float> selector);
+    double Count(Func<TSource, double> selector);
+    decimal Count(Func<TSource, decimal> selector);
+
     int Avg(Func<TSource, int?> selector);
     long Avg(Func<TSource, long?> selector);
     decimal Avg(Func<TSource, decimal?> selector);
@@ -17,9 +24,9 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
     decimal Avg(Func<TSource, double?> selector);
     int Avg(Func<TSource, int> selector);
     long Avg(Func<TSource, long> selector);
-    decimal Avg(Func<TSource, decimal> selector);
     decimal Avg(Func<TSource, float> selector);
     decimal Avg(Func<TSource, double> selector);
+    decimal Avg(Func<TSource, decimal> selector);
     
     int Min(Func<TSource, int?> selector);
     long Min(Func<TSource, long?> selector);
@@ -28,9 +35,9 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
     decimal Min(Func<TSource, double?> selector);
     int Min(Func<TSource, int> selector);
     long Min(Func<TSource, long> selector);
-    decimal Min(Func<TSource, decimal> selector);
     decimal Min(Func<TSource, float> selector);
     decimal Min(Func<TSource, double> selector);
+    decimal Min(Func<TSource, decimal> selector);
     
     int Max(Func<TSource, int?> selector);
     long Max(Func<TSource, long?> selector);
@@ -39,9 +46,9 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
     decimal Max(Func<TSource, double?> selector);
     int Max(Func<TSource, int> selector);
     long Max(Func<TSource, long> selector);
-    decimal Max(Func<TSource, decimal> selector);
     decimal Max(Func<TSource, float> selector);
     decimal Max(Func<TSource, double> selector);
+    decimal Max(Func<TSource, decimal> selector);
     
     int Sum(Func<TSource, int?> selector);
     long Sum(Func<TSource, long?> selector);
@@ -50,9 +57,9 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
     decimal Sum(Func<TSource, double?> selector);
     int Sum(Func<TSource, int> selector);
     long Sum(Func<TSource, long> selector);
-    decimal Sum(Func<TSource, decimal> selector);
     decimal Sum(Func<TSource, float> selector);
     decimal Sum(Func<TSource, double> selector);
+    decimal Sum(Func<TSource, decimal> selector);
 
     short[] TopK(Func<TSource, short> selector, int k);
     int[] TopK(Func<TSource, int> selector, int k);
@@ -71,29 +78,29 @@ namespace Kafka.DotNet.ksqlDB.KSql.Linq
     string EarliestByOffset(Func<TSource, string> selector);
     int EarliestByOffset(Func<TSource, int> selector);
     long EarliestByOffset(Func<TSource, long> selector);
-    decimal EarliestByOffset(Func<TSource, decimal> selector);
     float EarliestByOffset(Func<TSource, float> selector);
     double EarliestByOffset(Func<TSource, double> selector);
+    decimal EarliestByOffset(Func<TSource, decimal> selector);
 
     string EarliestByOffsetAllowNulls(Func<TSource, string> selector);
     int? EarliestByOffsetAllowNulls(Func<TSource, int?> selector);
     long? EarliestByOffsetAllowNulls(Func<TSource, long?> selector);
-    decimal? EarliestByOffsetAllowNulls(Func<TSource, decimal?> selector);
     float? EarliestByOffsetAllowNulls(Func<TSource, float?> selector);
     double? EarliestByOffsetAllowNulls(Func<TSource, double?> selector);
+    decimal? EarliestByOffsetAllowNulls(Func<TSource, decimal?> selector);
     
     string LatestByOffset(Func<TSource, string> selector);
     int LatestByOffset(Func<TSource, int> selector);
     long LatestByOffset(Func<TSource, long> selector);
-    decimal LatestByOffset(Func<TSource, decimal> selector);
     float LatestByOffset(Func<TSource, float> selector);
     double LatestByOffset(Func<TSource, double> selector);
+    decimal LatestByOffset(Func<TSource, decimal> selector);
 
     string LatestByOffsetAllowNulls(Func<TSource, string> selector);
     int? LatestByOffsetAllowNulls(Func<TSource, int?> selector);
     long? LatestByOffsetAllowNulls(Func<TSource, long?> selector);
-    decimal? LatestByOffsetAllowNulls(Func<TSource, decimal?> selector);
     float? LatestByOffsetAllowNulls(Func<TSource, float?> selector);
     double? LatestByOffsetAllowNulls(Func<TSource, double?> selector);
+    decimal? LatestByOffsetAllowNulls(Func<TSource, decimal?> selector);
   }
 }